chore: split device & driver controls files
